Move product.js in to run.js
